requirements.txt
alembic==1.4.3

asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
blinker==1.4

certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3

confusable-homoglyphs==3.2.0
cryptography==2.8

dbus-python==1.2.16

distro==1.4.0

dj-database-url==0.5.0
Django==3.1.6
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==3.2.0
django-utils-six==2.0
djangorestframework==3.12.2
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
entrypoints==0.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Reuploaded==0.3.2
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.4
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==2.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2


launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3

macaroonbakery==1.3.1
Mako==1.1.3
markdown2==2.3.10
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==5.0.0
protobuf==3.6.1
psycopg2==2.7.4
psycopg2-binary==2.8.6
pycairo==1.16.2
pycups==1.9.73

PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyperclip==1.8.1


python-dateutil==2.7.3

python-decouple==3.4
python-editor==1.0.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.1
SQLAlchemy==1.3.20
sqlparse==0.4.1

toml==0.10.2
typed-ast==1.4.1


ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
visitor==0.1.3
wadllib==1.3.3
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
zipp==3.4.0

apturl==0.5.2
Brlapi==0.7.0
command-not-found==0.3
cupshelpers==1.0
defer==1.0.6
distro-info===0.23ubuntu1
keyring==18.0.1
language-selector==0.1
louis==3.12.0
PyGObject==3.36.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-debian===0.1.36ubuntu1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
